🚸 Set PORTAGE_CONFIGROOT to / by default
In case gcm uses portage installed in a virtual environment,
portage looks for configuration files where it's installed.
This behavior might be confusing, so it's better to make
portage use configuration files from the root by default.

For a better understanding, see the descriptions of the
EPREFIX and PORTAGE_CONFIGROOT environment variables:
  - `man emerge`
  - https://dev.gentoo.org/~zmedico/portage/doc/man/emerge.1.html